+++ features/call/src/main/kotlin/io/element/android/features/call/CallIntentDataParser.kt
@@ -21,13 +21,13 @@ import javax.inject.Inject
 
 class CallIntentDataParser @Inject constructor() {
 
-    private val validHttpSchemes = sequenceOf("http", "https")
+    private val validHttpSchemes = sequenceOf("https")
 
     fun parse(data: String?): String? {
         val parsedUrl = data?.let { Uri.parse(data) } ?: return null
         val scheme = parsedUrl.scheme
         return when {
-            scheme in validHttpSchemes && parsedUrl.host == "call.element.io" -> data
+            scheme in validHttpSchemes && parsedUrl.host == "call.element.io" -> parsedUrl
             scheme == "element" && parsedUrl.host == "call" -> {
                 // We use this custom scheme to load arbitrary URLs for other instances of Element Call,
                 // so we can only verify it's an HTTP/HTTPs URL with a non-empty host
@@ -40,14 +40,36 @@ class CallIntentDataParser @Inject constructor() {
             }
             // This should never be possible, but we still need to take into account the possibility
             else -> null
-        }
+        }?.withCustomParameters()
     }
 
-    private fun Uri.getUrlParameter(): String? {
+    private fun Uri.getUrlParameter(): Uri? {
         return getQueryParameter("url")
-            ?.takeIf {
-                val internalUri = Uri.parse(it)
-                internalUri.scheme in validHttpSchemes && !internalUri.host.isNullOrBlank()
+            ?.let { urlParameter ->
+                Uri.parse(urlParameter).takeIf { uri ->
+                    uri.scheme in validHttpSchemes && !uri.host.isNullOrBlank()
+                }
             }
     }
 }
+
+/**
+ * Ensure the uri has the following parameters and value:
+ * - appPrompt=false
+ * - confineToRoom=true
+ * to ensure that the rendering will bo correct on the embedded Webview.
+ */
+private fun Uri.withCustomParameters(): String {
+    val builder = buildUpon()
+    builder.clearQuery()
+    queryParameterNames.forEach {
+        if (it == APP_PROMPT_PARAMETER || it == CONFINE_TO_ROOM_PARAMETER) return@forEach
+        builder.appendQueryParameter(it, getQueryParameter(it))
+    }
+    builder.appendQueryParameter(APP_PROMPT_PARAMETER, "false")
+    builder.appendQueryParameter(CONFINE_TO_ROOM_PARAMETER, "true")
+    return builder.build().toString()
+}
+
+private const val APP_PROMPT_PARAMETER = "appPrompt"
+private const val CONFINE_TO_ROOM_PARAMETER = "confineToRoom"